requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.8.0
mysql-connector-python==8.1.0
openai==1.51.2
numpy==1.24.3
gunicorn==21.2.0
python-dotenv==1.0.0